style: run prettier and eslint

+++ packages/demo-dapp-election/contracts/Election.sol
@@ -3,35 +3,33 @@ pragma solidity ^0.7.0;
 contract Election {
     // Model a Candidate
     struct Candidate {
-        uint id;
+        uint256 id;
         string name;
-        uint voteCount;
+        uint256 voteCount;
     }
 
     // Store accounts that have voted
     mapping(address => bool) public voters;
     // Store Candidates
     // Fetch Candidate
-    mapping(uint => Candidate) public candidates;
+    mapping(uint256 => Candidate) public candidates;
     // Store Candidates Count
-    uint public candidatesCount;
+    uint256 public candidatesCount;
 
     // voted event
-    event votedEvent (
-        uint indexed _candidateId
-    );
+    event votedEvent(uint256 indexed _candidateId);
 
-    constructor () public {
+    constructor() public {
         addCandidate("Candidate 1");
         addCandidate("Candidate 2");
     }
 
-    function addCandidate (string memory _name) private {
-        candidatesCount ++;
+    function addCandidate(string memory _name) private {
+        candidatesCount++;
         candidates[candidatesCount] = Candidate(candidatesCount, _name, 0);
     }
 
-    function vote (uint _candidateId) public {
+    function vote(uint256 _candidateId) public {
         // require that they haven't voted before
         require(!voters[msg.sender]);
 
@@ -42,7 +40,7 @@ contract Election {
         voters[msg.sender] = true;
 
         // update candidate vote Count
-        candidates[_candidateId].voteCount ++;
+        candidates[_candidateId].voteCount++;
 
         // trigger voted event
         emit votedEvent(_candidateId);
